🚧 Rename MovieInformationContainer to MovieInformation
