+++ src/components/MovieInformation.js
